[mlir] add ComplexTypeConstraints to DDR, allowing matching of type parameters
This contains a bit of an ugly hack for dealing with the fact that the SymbolInfo for matched type parmeters needs to keep track of the original op, the opArgIndex of the oprand that's constrained (as usual) but then also the index of the type parameter, which doesn't work with the "DagAndConstant" system used by SymbolInfo, which can only track one index.

The current workaround just stores them as opArgIndex * 1000 + paramIndex and recovers them when needed.